+++ components/mailbox/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - [Mailbox] Added `onError` custom event when an error occurs [#262](https://github.com/nylas/components/pull/262)
 
+- Added ability to reply to email/thread [#228](https://github.com/nylas/components/issues/228)
+
 ## Bug Fixes
 
 # v1.1.5 (2021-12-10)